Built the EXE - Seems We Have a Problem?
Rebuilt the EXE so it's on the most recent version. With that said, we appear to have a problem. Some of the variables seem to be sharing values, namely the integers. I don't know why this happens, as all of them are different variables. Because of this, the data is not displaying correctly. To get around this, I'll need screenshots of the application variables instead. I will need to make a note of this later.