Figured Out the Problem
I figured out the problem with GameSceneA and B and being able to transition properly. I made these scenes by duplicating existing scenes. For some reason when I did that it screwed something up, causing the scene switch to fail when done through an executed version of the game. Making a new scene and copying everything over seems to fix that, so I'm going to do that now.