The variable "shotsHitCounts" does not work, so I'm ignoring it.
For some reason, this specific list variable causes the EXE to fail. All the other lists work fine, but this specific one causes a failure. I don't know why this specific variable causes a failure, but it does. I put in a try-catch block so that it ignores this variable right now. Later down the line this should be fixed.